feat: implement QueryParams class with methods for handling URL query parameters
